General refactoring
Various changes to make the program more robust:
Change Development int properties to uints
Change Development Cost to TokenCollection
Change some functions to expression bodies
Change GameDataLoader to be in line with refactror
Simplify GameDataLoader methods
Change Noble int properties to uints
Clean up Player
Add comparison logic for TokenCollection
Update tests to align with refactor

+++ splendor-lib/Player.cs
@@ -21,7 +21,7 @@ namespace splendor_lib
             _noblesInternal = new List<Noble>();
             _reservedDevelopmentsInternal = new List<Development>(3);
         }
-        public int Prestige => _purchasedDevelopmentsInternal.Sum(d => d.Prestige) + _noblesInternal.Sum(n => n.Prestige);
+        public uint Prestige => (uint)(_purchasedDevelopmentsInternal.Sum(d => d.Prestige) + _noblesInternal.Sum(n => n.Prestige));
         public bool HasTooManyTokens => _tokensInternal.TotalTokens > 10;
         public bool HandFull => _reservedDevelopmentsInternal.Count == 3;
         public void TakeNoble(Noble noble) => _noblesInternal.Add(noble);
@@ -63,7 +63,6 @@ namespace splendor_lib
 
             return true;
         }
-
         public bool TryPay(TokenCollection price)
         {
             if (!CanPay(price)) return false;
